Adjust FailedDelayedJobs alert
Using the delta function over a 10 minute period means that Alert
Manager will send a message saying the issue has been resolved even
though there is still a failed submission in the queue.

Using avg instead of the delta should result in the resolved alert not
being sent until the number of failed submissions in the queue has
dropped back down to 0.