Send non prod alerts to low severity slack channel
The #form-builder-alerts channel was a mixture of alerts from all 4
environments. This generated a lot of noise.

A new slack channel, #form-builder-alerts-low-severity, was created to
receive all non production alerts instead.

In [this ticket](https://github.com/ministryofjustice/cloud-platform/issues/1858)
Cloud Platform created the necessary plumbing for these alerts to be
sent to the correct channel.

`live-production` alerts should remain being sent to
the #form-builder-alerts channel.