Simplify circle config
Using a matrix to repeat the different jobs with different parameters.

Deleted the publisher prometheus rules as it is no longer used and also we are not getting any alerts. The already applied rules will remain in the namespace anyways unless we manually delete their manifests.

Note this probably will fail to deploy as we need to change the way the credentials are obtained. This is just a first pass.