Update circle config and scripts (#45)
* Simplify circle config

Using a matrix to repeat the different jobs with different parameters.

Deleted the publisher prometheus rules as it is no longer used and also we are not getting any alerts. The already applied rules will remain in the namespace anyways unless we manually delete their manifests.

Note this probably will fail to deploy as we need to change the way the credentials are obtained. This is just a first pass.

* Use version 2.1

* Using contexts

* Use ruby docker image

* Tweaks

* Apply grafana job

* Apply prometheus

* Mystery with base64 decoding

* Remove spaces

* Add services back

* Only use `moj-forms` context

* Revert "Only use `moj-forms` context"

This reverts commit 95bf2ce3cd59dd085881f8a56fb9321fd84d2fc0.

* Uncomment grafana job

* Uncomment hmcts job

* Add license and codeowners files

* Remove branch from circle config